+++ modules/web/src/app/dynamic/enterprise/cluster-backups/list/restore/template.html
@@ -132,7 +132,7 @@ limitations under the License.
                class="km-table-actions">
             <button mat-icon-button
                     matTooltip="Delete Restore"
-                    (click)="deleteRestore(element.name);">
+                    (click)="deleteRestores([element]);">
               <i class="km-icon-mask km-icon-delete"></i>
             </button>
           </div>
